chore: renames to properties-explorer
