+++ app/src/app/properties-explorer/PropertiesExplorer.module.scss.d.ts
@@ -1,0 +1,16 @@
+export type Styles = {
+  "properties-explorer": string;
+  "z-depth-0": string;
+  "z-depth-1": string;
+  "z-depth-1-half": string;
+  "z-depth-2": string;
+  "z-depth-3": string;
+  "z-depth-4": string;
+  "z-depth-5": string;
+};
+
+export type ClassNames = keyof Styles;
+
+declare const styles: Styles;
+
+export default styles;
