+++ app/src/app/properties-explorer/PropertiesIndex/PropertiesIndex.module.scss.d.ts
