+++ app/src/app/properties-explorer/PropertiesIndex/PropertiesIndex.test.tsx
