+++ app/src/app/properties-explorer/PropertiesIndex/PropertiesIndex.tsx
