+++ app/src/app/properties-explorer/PropertiesIndex/PropertiesIndex.types.ts
