+++ app/src/app/properties-explorer/PropertiesIndex/PropertiesIndexContainer.tsx
