+++ app/src/app/properties-explorer/property-card/PropertyCard.module.scss
