+++ app/src/app/properties-explorer/property-card/PropertyCard.test.tsx
