+++ app/src/app/properties-explorer/property-card/PropertyCard.types.ts
