+++ app/src/app/property-details/PropertyDetailsContainer.tsx
@@ -13,8 +13,8 @@ import {
 import { ConditionalEscrowMethods, ConditionalEscrowValues } from "providers/near/contract/conditional-escrow.types";
 import { Typography } from "ui/typography/Typography";
 import ipfs from "providers/ipfs";
-import { DEFAULT_PROPERTY_CARD_PROPS } from "app/properties-index/property-card/PropertyCard";
-import { PropertyCardProps } from "app/properties-index/property-card/PropertyCard.types";
+import { DEFAULT_PROPERTY_CARD_PROPS } from "app/properties-explorer/property-card/PropertyCard";
+import { PropertyCardProps } from "app/properties-explorer/property-card/PropertyCard.types";
 
 import { PropertyDetails2 } from "./PropertyDetails2";
